move method channel impl to subfolder
