+++ lib/gigya_flutter_plugin.dart
@@ -1,14 +1,14 @@
 import 'src/models/enums/social_provider.dart';
 import 'src/models/screenset_event.dart';
 import 'src/platform_interface/gigya_flutter_plugin_platform_interface.dart';
-import 'src/services/interruption_resolver/interruption_resolver.dart';
-import 'src/services/otp_service/otp_service.dart';
-import 'src/services/web_authentication_service/web_authentication_service.dart';
+import 'src/services/interruption_resolver.dart';
+import 'src/services/otp_service.dart';
+import 'src/services/web_authentication_service.dart';
 
 export 'src/models/account.dart';
 export 'src/models/address.dart';
 export 'src/models/certification.dart';
-export 'src/models/conflicting_accounts.dart';
+export 'src/models/conflicting_account.dart';
 export 'src/models/education.dart';
 export 'src/models/emails.dart';
 export 'src/models/enums/screen_set_event_type.dart';
@@ -26,9 +26,9 @@ export 'src/models/screenset_event.dart';
 export 'src/models/session_info.dart';
 export 'src/models/skill.dart';
 export 'src/models/work.dart';
-export 'src/services/interruption_resolver/interruption_resolver.dart';
-export 'src/services/otp_service/otp_service.dart';
-export 'src/services/web_authentication_service/web_authentication_service.dart';
+export 'src/services/interruption_resolver.dart';
+export 'src/services/otp_service.dart';
+export 'src/services/web_authentication_service.dart';
 
 /// This class represents the Gigya SDK plugin.
 class GigyaSdk {
