+++ lib/src/method_channel/gigya_flutter_plugin_method_channel.dart
@@ -7,13 +7,13 @@ import '../models/enums/methods.dart';
 import '../models/enums/social_provider.dart';
 import '../models/gigya_error.dart';
 import '../models/screenset_event.dart';
-import '../services/interruption_resolver/interruption_resolver.dart';
-import '../services/interruption_resolver/method_channel_interruption_resolver.dart';
-import '../services/otp_service/method_channel_otp_service.dart';
-import '../services/otp_service/otp_service.dart';
-import '../services/web_authentication_service/method_channel_web_authentication_service.dart';
-import '../services/web_authentication_service/web_authentication_service.dart';
-import 'gigya_flutter_plugin_platform_interface.dart';
+import '../platform_interface/gigya_flutter_plugin_platform_interface.dart';
+import '../services/interruption_resolver.dart';
+import '../services/otp_service.dart';
+import '../services/web_authentication_service.dart';
+import 'method_channel_interruption_resolver.dart';
+import 'method_channel_otp_service.dart';
+import 'method_channel_web_authentication_service.dart';
 
 /// An implementation of [GigyaFlutterPluginPlatform] that uses method channels.
 class MethodChannelGigyaFlutterPlugin extends GigyaFlutterPluginPlatform {
