+++ lib/src/platform_interface/gigya_flutter_plugin_platform_interface.dart
@@ -1,11 +1,11 @@
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
+import '../method_channel/gigya_flutter_plugin_method_channel.dart';
 import '../models/enums/social_provider.dart';
 import '../models/screenset_event.dart';
-import '../services/interruption_resolver/interruption_resolver.dart';
-import '../services/otp_service/otp_service.dart';
-import '../services/web_authentication_service/web_authentication_service.dart';
-import 'gigya_flutter_plugin_method_channel.dart';
+import '../services/interruption_resolver.dart';
+import '../services/otp_service.dart';
+import '../services/web_authentication_service.dart';
 
 /// The platform interface for the Gigya Flutter Plugin.
 abstract class GigyaFlutterPluginPlatform extends PlatformInterface {
