+++ lib/src/services/web_authentication_service.dart
